# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple SMS gateway with balance management and operator failover.",
        "title": "SMS Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/balance": {
            "get": {
                "description": "Returns current balance and transaction history",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Get user balance and transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "user_id is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/balance/add": {
            "post": {
                "description": "Increases user balance and records transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Add balance for user",
                "parameters": [
                    {
                        "description": "Add balance request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/balance.AddBalancePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "done",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sms/history": {
            "get": {
                "description": "Returns sent SMS history for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "Get SMS history for user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending|sending|done|failed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by sms_identifier",
                        "name": "sms_identifier",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "user_id is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sms/send": {
            "post": {
                "description": "Deducts balance, enqueues SMS for processing, returns processing ack",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms"
                ],
                "summary": "Send SMS request",
                "parameters": [
                    {
                        "description": "SMS request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SMS"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ack with sms_identifier",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "dont have Not Enough Balance",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "balance.AddBalancePayload": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.SMS": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "integer"
                },
                "recipients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sms_identifier": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.Type"
                }
            }
        },
        "model.Type": {
            "type": "string",
            "enum": [
                "normal",
                "express"
            ],
            "x-enum-varnames": [
                "NORMAL",
                "EXPRESS"
            ]
        }
    }
}